requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
blinker==1.7.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama
cryptography==41.0.5
dataclasses-json==0.6.2
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.103.1
fastapi-mail==1.4.1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
huggingface_hub
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
llama-cpp-python
langchain==0.0.334
langsmith==0.0.63
loguru==0.7.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
passlib==1.7.4
pinecone-client==2.2.4
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyMySQL==1.1.0
pypdf
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
sentence-transformers
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
starlette==0.27.0
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.9.2
